address PR review comments: move caches to subpackages and remove unused Update methods
