+++ pkg/controller/core/workload_controller.go
@@ -1690,14 +1690,10 @@ func (h *resourceUpdatesHandler) Update(ctx context.Context, e event.UpdateEvent
 	ctx = ctrl.LoggerInto(ctx, log)
 	log.V(5).Info("Update event")
 	if lr, ok := e.ObjectNew.(*corev1.LimitRange); ok {
-		if oldLr, ok := e.ObjectOld.(*corev1.LimitRange); ok {
-			h.r.queues.LimitRangeCache().Update(oldLr, lr)
-		}
+		h.r.queues.LimitRangeCache().AddOrUpdate(lr)
 	}
 	if rc, ok := e.ObjectNew.(*nodev1.RuntimeClass); ok {
-		if oldRc, ok := e.ObjectOld.(*nodev1.RuntimeClass); ok {
-			h.r.queues.RuntimeClassCache().Update(oldRc, rc)
-		}
+		h.r.queues.RuntimeClassCache().Add(rc)
 	}
 	h.handle(ctx, e.ObjectNew, q)
 	if oldLr, isLr := e.ObjectOld.(*corev1.LimitRange); isLr {
